third_party/blink/web_tests/wpt_internal/prerender/clients-matchall.https.html

<!DOCTYPE html>
<title>Service Worker: Clients.matchAll with a prerender page</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

async function unregisterServiceWorker(scope) {
  const registration = await navigator.serviceWorker.getRegistration(scope);
  if (!registration)
    return;
  return registration.unregister();
}

function getNewestWorker(registration) {
  if (registration.installing)
    return registration.installing;
  if (registration.waiting)
    return registration.waiting;
  if (registration.active)
    return registration.active;
}

promise_test(async t => {
  const workerUrl = 'resources/clients-matchall-service-worker.js';
  const pageUrl = 'resources/prerendered-page.html';

  // Start a service worker.
  await unregisterServiceWorker(workerUrl);
  const registration = await navigator.serviceWorker.register(workerUrl);
  t.add_cleanup(_ => registration.unregister());

  // Observe the message from a prerendered page.
  const bc = new BroadcastChannel('prerender-channel');
  t.add_cleanup(_ => bc.close());
  const messagePromise = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {once:true});
  });

  startPrerendering(pageUrl);

  const result = await messagePromise;
  assert_equals(result, 'prerender success');

  // The service worker returns the list of client urls exposed by
  // Clients#matchAll().
  const message = await new Promise(resolve => {
    navigator.serviceWorker.onmessage = resolve;
    getNewestWorker(registration).postMessage('invoke clients.matchAll()');
  });

  assert_array_equals(
    message.data,
    [ window.location.href, new URL(pageUrl, location).toString() ]);
}, 'The client urls (including a prerender page) are exposed by ' +
   'Clients#matchAll()');

</script>
</html>
